home work 6 fix

+++ app/src/main/java/ru/thstdio/aa2020/data/Genre.kt
@@ -1,3 +1,3 @@
 package ru.thstdio.aa2020.data
 
-data class Genre(val id: Int, val name: String)
+data class Genre(val id: Long, val name: String)
