+++ app/src/main/java/ru/thstdio/aa2020/ui/Navigation.kt
@@ -1,6 +1,6 @@
 package ru.thstdio.aa2020.ui
 
 interface Navigation {
-    fun openDetail(id: Int)
+    fun openDetail(id: Long)
     fun back()
 }
